[WIP] check map and add support for linux mlx

+++ src/types/coordinate.c
@@ -35,7 +35,7 @@ t_coord_f64	set_f64(float x, float y)
 	return ((t_coord_f64){x, y});
 }
 
-t_coord_i32	set_i32(float x, float y)
+t_coord_i32	set_i32(int32_t x, int32_t y)
 {
 	return ((t_coord_i32){x, y});
 }
