+++ src/types/cube_types.h
@@ -13,6 +13,9 @@
 #ifndef CUBE_TYPES_H
 # define CUBE_TYPES_H
 
+/* int32_t */
+#include <stdint.h>
+
 /* ************************************************************************** */
 
 typedef struct s_coord_i32 {
@@ -31,7 +34,7 @@ t_coord_i32	init_i32(void);
 t_coord_f64	init_f64(void);
 
 t_coord_f64	set_f64(float x, float y);
-t_coord_i32	set_i32(float x, float y);
+t_coord_i32	set_i32(int32_t x, int32_t y);
 
 t_coord_f64	add_f64(t_coord_f64 a, t_coord_f64 b);
